requirements.txt
# Core dependencies
httpx==0.25.0
python-dotenv==1.0.0
pydantic==2.5.0

# Database
asyncpg==0.29.0
sqlalchemy==2.0.23

# Async support
aiohttp==3.9.0

# CLI tools
rich==13.7.0
typer==0.9.0

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
